pyproject.toml
[project]
name = "reservacia"
version = "0.1.0"
description = "Single-tenant restaurant reservation API for an AI agent consumer."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyjwt>=2.8",
    "httpx>=0.26",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
